requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
idna==3.3
loguru==0.6.0
peewee==3.14.9
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot-calendar==1.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
telebot-calendar==1.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
win32-setctime==1.1.0
